+++ source/main/modules/fsal/fsal-file.ts
@@ -243,8 +243,7 @@ export async function parse (filePath: string, cache: FSALCache|null, parent: Di
     file.size = stat.size
   } catch (err: any) {
     global.log.error('Error reading file ' + filePath, err)
-    // Re-throw a nicer and more meaningful message
-    throw new Error(`Could not read file ${filePath}: ${String(err.message)}`)
+    throw err // Re-throw
   }
 
   // Before reading in the full file and parsing it,